requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
firebase-admin==6.4.0
flask-cors==4.0.0
python-dotenv==1.0.0
Flask-Login==0.6.3

